+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/GenericJigsawStructure.java
@@ -1,6 +1,5 @@
 package com.telepathicgrunt.repurposedstructures.world.structures;
 
-import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.modinit.RSStructureTagMap;
 import com.telepathicgrunt.repurposedstructures.modinit.RSStructures;
 import net.minecraft.util.ResourceLocation;
@@ -14,8 +13,8 @@ import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.MobSpawnInfo;
 import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.ChunkGenerator;
+import net.minecraft.world.gen.Heightmap;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
-import net.minecraft.world.gen.feature.StructureFeature;
 import net.minecraft.world.gen.feature.jigsaw.JigsawManager;
 import net.minecraft.world.gen.feature.structure.*;
 import net.minecraft.world.gen.feature.template.TemplateManager;
@@ -26,28 +25,55 @@ import java.util.List;
 import java.util.Set;
 
 public class GenericJigsawStructure extends AbstractBaseStructure {
-    private final ResourceLocation START_POOL;
-    private final int STRUCTURE_SIZE;
-    private final int CENTER_OFFSET;
-    private final int BIOME_RANGE;
-    private final int STRUCTURE_BLACKLIST_RANGE;
-    private final Set<RSStructureTagMap.STRUCTURE_TAGS> AVOID_STRUCTURES_SET;
-    private final List<MobSpawnInfo.Spawners> MONSTER_SPAWNS;
-    private final List<MobSpawnInfo.Spawners> CREATURE_SPAWNS;
+    private final ResourceLocation startPool;
+    private final int structureSize;
+    private final int centerOffset;
+    private final int biomeRange;
+    private final int structureBlacklistRange;
+    private final Set<RSStructureTagMap.STRUCTURE_TAGS> structureTagsSet;
+    private final List<MobSpawnInfo.Spawners> monsterSpawns;
+    private final List<MobSpawnInfo.Spawners> creatureSpawns;
+    private final int allowTerrainHeightRange;
+    private final int terrainHeightRadius;
+
 
     public GenericJigsawStructure(ResourceLocation poolRL, int structureSize, int centerOffset,
                                   int biomeRange, int structureBlacklistRange, Set<RSStructureTagMap.STRUCTURE_TAGS> avoidStructuresSet)
     {
         super(NoFeatureConfig.CODEC);
-        START_POOL = poolRL;
-        STRUCTURE_SIZE = structureSize;
-        CENTER_OFFSET = centerOffset;
-        BIOME_RANGE = biomeRange;
-        STRUCTURE_BLACKLIST_RANGE = structureBlacklistRange;
-        AVOID_STRUCTURES_SET = avoidStructuresSet;
-        RSStructures.RS_STRUCTURE_START_PIECES.add(START_POOL);
-        MONSTER_SPAWNS = new ArrayList<>();
-        CREATURE_SPAWNS = new ArrayList<>();
+
+        this.startPool = poolRL;
+        this.structureSize = structureSize;
+        this.centerOffset = centerOffset;
+        this.biomeRange = biomeRange;
+        this.structureBlacklistRange = structureBlacklistRange;
+        this.structureTagsSet = avoidStructuresSet;
+        this.monsterSpawns = new ArrayList<>();
+        this.creatureSpawns = new ArrayList<>();
+        this.allowTerrainHeightRange = -1;
+        this.terrainHeightRadius = 0;
+
+        RSStructures.RS_STRUCTURE_START_PIECES.add(startPool);
+    }
+
+    public GenericJigsawStructure(ResourceLocation poolRL, int structureSize, int centerOffset,
+                                  int biomeRange, int structureBlacklistRange, Set<RSStructureTagMap.STRUCTURE_TAGS> avoidStructuresSet,
+                                  int allowTerrainHeightRange, int terrainHeightRadius)
+    {
+        super(NoFeatureConfig.CODEC);
+
+        this.startPool = poolRL;
+        this.structureSize = structureSize;
+        this.centerOffset = centerOffset;
+        this.biomeRange = biomeRange;
+        this.structureBlacklistRange = structureBlacklistRange;
+        this.structureTagsSet = avoidStructuresSet;
+        this.monsterSpawns = new ArrayList<>();
+        this.creatureSpawns = new ArrayList<>();
+        this.allowTerrainHeightRange = allowTerrainHeightRange;
+        this.terrainHeightRadius = terrainHeightRadius;
+
+        RSStructures.RS_STRUCTURE_START_PIECES.add(startPool);
     }
 
     public GenericJigsawStructure(ResourceLocation poolID, int structureSize, int centerOffset,
@@ -55,31 +81,35 @@ public class GenericJigsawStructure extends AbstractBaseStructure {
                                   List<MobSpawnInfo.Spawners> monster_spawns, List<MobSpawnInfo.Spawners> creature_spawns)
     {
         super(NoFeatureConfig.CODEC);
-        START_POOL = poolID;
-        STRUCTURE_SIZE = structureSize;
-        CENTER_OFFSET = centerOffset;
-        BIOME_RANGE = biomeRange;
-        STRUCTURE_BLACKLIST_RANGE = structureBlacklistRange;
-        AVOID_STRUCTURES_SET = avoidStructuresSet;
-        RSStructures.RS_STRUCTURE_START_PIECES.add(START_POOL);
-        MONSTER_SPAWNS = monster_spawns;
-        CREATURE_SPAWNS = creature_spawns;
+
+        this.startPool = poolID;
+        this.structureSize = structureSize;
+        this.centerOffset = centerOffset;
+        this.biomeRange = biomeRange;
+        this.structureBlacklistRange = structureBlacklistRange;
+        this.structureTagsSet = avoidStructuresSet;
+        this.monsterSpawns = monster_spawns;
+        this.creatureSpawns = creature_spawns;
+        this. allowTerrainHeightRange = -1;
+        this.terrainHeightRadius = 0;
+
+        RSStructures.RS_STRUCTURE_START_PIECES.add(startPool);
     }
 
     @Override
     public List<MobSpawnInfo.Spawners> getDefaultSpawnList() {
-        return MONSTER_SPAWNS;
+        return monsterSpawns;
     }
 
     @Override
     public List<MobSpawnInfo.Spawners> getDefaultCreatureSpawnList() {
-        return CREATURE_SPAWNS;
+        return creatureSpawns;
     }
 
     @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeProvider biomeSource, long seed, SharedSeedRandom chunkRandom, int chunkX, int chunkZ, Biome biome, ChunkPos chunkPos, NoFeatureConfig NoFeatureConfig) {
-         for (int curChunkX = chunkX - BIOME_RANGE; curChunkX <= chunkX + BIOME_RANGE; curChunkX++) {
-            for (int curChunkZ = chunkZ - BIOME_RANGE; curChunkZ <= chunkZ + BIOME_RANGE; curChunkZ++) {
+         for (int curChunkX = chunkX - biomeRange; curChunkX <= chunkX + biomeRange; curChunkX++) {
+            for (int curChunkZ = chunkZ - biomeRange; curChunkZ <= chunkZ + biomeRange; curChunkZ++) {
                 if (!biomeSource.getBiomeForNoiseGen(curChunkX << 2, 60, curChunkZ << 2).getGenerationSettings().hasStructureFeature(this)) {
                     return false;
                 }
@@ -87,9 +117,9 @@ public class GenericJigsawStructure extends AbstractBaseStructure {
         }
 
         //cannot be near other specified structure
-        for (int curChunkX = chunkX - STRUCTURE_BLACKLIST_RANGE; curChunkX <= chunkX + STRUCTURE_BLACKLIST_RANGE; curChunkX++) {
-            for (int curChunkZ = chunkZ - STRUCTURE_BLACKLIST_RANGE; curChunkZ <= chunkZ + STRUCTURE_BLACKLIST_RANGE; curChunkZ++) {
-                for(RSStructureTagMap.STRUCTURE_TAGS tag : AVOID_STRUCTURES_SET){
+        for (int curChunkX = chunkX - structureBlacklistRange; curChunkX <= chunkX + structureBlacklistRange; curChunkX++) {
+            for (int curChunkZ = chunkZ - structureBlacklistRange; curChunkZ <= chunkZ + structureBlacklistRange; curChunkZ++) {
+                for(RSStructureTagMap.STRUCTURE_TAGS tag : structureTagsSet){
                     for(Structure<?> structureFeature : RSStructureTagMap.REVERSED_TAGGED_STRUCTURES.get(tag)){
                         StructureSeparationSettings structureConfig = chunkGenerator.getStructuresConfig().getForType(structureFeature);
                         if(structureConfig != null){
@@ -101,6 +131,21 @@ public class GenericJigsawStructure extends AbstractBaseStructure {
                     }
                 }
             }
+        }
+
+        if(allowTerrainHeightRange != -1){
+            int maxTerrainHeight = Integer.MIN_VALUE;
+            int minTerrainHeight = Integer.MAX_VALUE;
+
+            for (int curChunkX = chunkX - terrainHeightRadius; curChunkX <= chunkX + terrainHeightRadius; curChunkX++) {
+                for (int curChunkZ = chunkZ - terrainHeightRadius; curChunkZ <= chunkZ + terrainHeightRadius; curChunkZ++) {
+                    int height = chunkGenerator.func_222531_c((curChunkX << 4) + 7, (curChunkZ << 4) + 7, Heightmap.Type.WORLD_SURFACE_WG);
+                    maxTerrainHeight = Math.max(maxTerrainHeight, height);
+                    minTerrainHeight = Math.min(minTerrainHeight, height);
+                }
+            }
+
+            return maxTerrainHeight - minTerrainHeight <= allowTerrainHeightRange;
         }
 
         return true;
@@ -122,8 +167,8 @@ public class GenericJigsawStructure extends AbstractBaseStructure {
             JigsawManager.method_30419(
                     dynamicRegistryManager,
                     new VillageConfig(() -> dynamicRegistryManager.get(Registry.TEMPLATE_POOL_WORLDGEN)
-                            .getOrDefault(START_POOL),
-                            STRUCTURE_SIZE),
+                            .getOrDefault(startPool),
+                            structureSize),
                     AbstractVillagePiece::new,
                     chunkGenerator,
                     structureManager,
@@ -133,7 +178,7 @@ public class GenericJigsawStructure extends AbstractBaseStructure {
                     true,
                     true);
             this.recalculateStructureSize();
-            this.components.get(0).offset(0, CENTER_OFFSET, 0);
+            this.components.get(0).offset(0, centerOffset, 0);
         }
     }
 }